requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
coverage==4.4.1           # via pytest-cov
defusedxml==0.5.0         # via python3-openid
django-allauth==0.32.0
django-cors-middleware==1.3.1
django-helusers==0.4.1
django==1.11.3            # via django-allauth, django-helusers, drf-oidc-auth, dynamic-rest
djangorestframework==3.6.2
drf-oidc-auth==0.9        # via django-helusers
dynamic-rest==1.6.7
future==0.16.0            # via pyjwkest
idna==2.5                 # via requests
inflection==0.3.1         # via dynamic-rest
markdown==2.6.8
oauthlib==2.0.2           # via requests-oauthlib
py==1.4.34                # via pytest
pycryptodomex==3.4.6      # via pyjwkest
pyjwkest==1.3.6           # via drf-oidc-auth
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.1.3
python3-openid==3.1.0     # via django-allauth
pytz==2017.2              # via django
requests-cache==0.4.13
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.1
six==1.10.0               # via pyjwkest
urllib3==1.21.1           # via requests
